Move product-details to folder
