+++ web/app/containers/home/container.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import * as commands from '../../integration-manager/commands'
+import {fetchHomeData} from '../../integration-manager/home/commands'
 
 // Partials
 import HomeCarousel from './partials/home-carousel'
@@ -16,7 +16,7 @@ const Home = () => {
 }
 
 Home.fetcher = (url, dispatch) => {
-    dispatch(commands.fetchHomeData(url))
+    dispatch(fetchHomeData(url))
 }
 
 export default Home
