+++ web/app/integration-manager/_merlins-connector/commands.js
@@ -1,9 +1,6 @@
 import {makeFormEncodedRequest, makeJsonEncodedRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
-import {urlToPathKey} from '../../utils/utils'
 import {fetchPageData} from './utils'
 
-import {productDetailsParser} from '../../store/products/parser'
-import {pdpAddToCartFormParser} from './parsers'
 import {checkoutShippingParser, parseCheckoutData} from './checkout/parsers'
 import * as responses from './../responses'
 import {getCustomerEntityID} from '../../store/checkout/selectors'
@@ -11,20 +8,10 @@ import {getIsLoggedIn} from '../../containers/app/selectors'
 import {getShippingFormValues} from '../../store/form/selectors'
 import {getCart} from '../../store/cart/actions'
 import {browserHistory} from 'react-router'
-import {receiveFormInfo} from './../actions'
 import {removeAllNotifications} from '../../containers/app/actions'
 
 import * as homeCommands from './home/commands'
-
-export const fetchPdpData = (url) => (dispatch) => {
-    return dispatch(fetchPageData(url))
-        .then((res) => {
-            const [$, $response] = res
-            dispatch(responses.receivePdpProductData({[urlToPathKey(url)]: productDetailsParser($, $response)}))
-            dispatch(receiveFormInfo({[urlToPathKey(url)]: pdpAddToCartFormParser($, $response).formInfo}))
-        })
-        .catch((error) => { console.info(error.message) })
-}
+import * as productDetailsCommands from './product-details/commands'
 
 export const fetchCheckoutShippingData = (url) => (dispatch) => {
     return dispatch(fetchPageData(url))
@@ -155,12 +142,13 @@ export const submitSignIn = () => {
 }
 
 export default {
-    fetchPdpData,
     fetchCheckoutShippingData,
     addToCart,
     makeFormEncodedRequest,
     submitShipping,
     checkCustomerEmail,
     submitSignIn,
-    home: homeCommands
+
+    home: homeCommands,
+    productDetails: productDetailsCommands
 }
