+++ web/app/integration-manager/_merlins-connector/product-details/parsers.js
