+++ web/app/integration-manager/product-details/commands.js
@@ -1,0 +1,7 @@
+let connector = {}
+
+export const register = (commands) => {
+    connector = commands
+}
+
+export const fetchPdpData = (...args) => connector.fetchPdpData(...args)
