+++ web/app/main.jsx
@@ -19,7 +19,7 @@ import Stylesheet from './stylesheet.scss' // eslint-disable-line no-unused-vars
 import {analyticManager} from 'progressive-web-sdk/dist/analytics/analytic-manager'
 import {clientAnalytics} from './utils/analytics/client-analytics'
 
-import connector from './integration-manager/merlins-connector'
+import connector from './integration-manager/_merlins-connector'
 import {registerConnector} from './integration-manager'
 
 polyfill()
